backend/pyproject.toml
[tool.poetry]
name = "libre-pago"
version = "0.1.0"
description = "Backend API for LibrePago - Spreadsheets management"
authors = ["LibrePago Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.5"
pandas = "^3.0.1"
python-multipart = "^0.0.9"
redis = "^5.0.0"
python-json-logger = "^2.0.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
bcrypt = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["UP007", "B008", "E501"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.poetry.scripts]
start = "app.scripts.start:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
